+++ packages/discovery-react-components/src/components/SearchFacets/components/ShowMore/ShowMoreButton.tsx
